fix: error messages for schema operations

+++ java/registry/src/main/java/dev/sunbirdrc/registry/service/impl/RegistryServiceImpl.java
@@ -15,6 +15,7 @@ import dev.sunbirdrc.registry.dao.IRegistryDao;
 import dev.sunbirdrc.registry.dao.RegistryDaoImpl;
 import dev.sunbirdrc.registry.dao.VertexReader;
 import dev.sunbirdrc.registry.dao.VertexWriter;
+import dev.sunbirdrc.registry.exception.RecordNotFoundException;
 import dev.sunbirdrc.registry.exception.SignatureException;
 import dev.sunbirdrc.registry.middleware.util.Constants;
 import dev.sunbirdrc.registry.middleware.util.JSONUtil;
@@ -44,6 +45,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import static dev.sunbirdrc.registry.Constants.CREDENTIAL_TEMPLATE;
 import static dev.sunbirdrc.registry.Constants.Schema;
+import static dev.sunbirdrc.registry.exception.ErrorMessages.INVALID_ID_MESSAGE;
 
 @Service
 @Qualifier("sync")
@@ -150,6 +152,9 @@ public class RegistryServiceImpl implements RegistryService {
             ReadConfigurator configurator = ReadConfiguratorFactory.getOne(false);
             VertexReader vertexReader = new VertexReader(databaseProvider, graph, configurator, uuidPropertyName, definitionsManager);
             Vertex vertex = vertexReader.getVertex(null, uuid);
+            if(vertex == null) {
+                throw new RecordNotFoundException(INVALID_ID_MESSAGE);
+            }
             String index = vertex.property(Constants.TYPE_STR_JSON_LD).isPresent() ? (String) vertex.property(Constants.TYPE_STR_JSON_LD).value() : null;
             if (!StringUtils.isEmpty(index) && index.equals(Schema)) {
                 schemaService.deleteSchemaIfExists(vertex);
